[#30] 박람회 크롤링 API 형식 수정
